Add DMBS deployment
The PR adds Cinder in A/A on the Edge site
and Glance multistore. The central site
redeploy with updated configuration for
ceph cluster on the site and glance
multistore is required.
It's needed to create an raw image in
all the glance stores to test such
deployment with tempest. Unfortunately
openstack client does not support that
and tempest image does not include
qemu-img to covert the image to raw
format directly on the tempest pod.